Moved pixel shader processor types
