+++ src/ImageSharp/Processing/Extensions/PixelShaderExtensions.cs
@@ -1,7 +1,7 @@
 // Copyright (c) Six Labors and contributors.
 // Licensed under the Apache License, Version 2.0.
 
-using SixLabors.ImageSharp.Processing.Processors.PixelShading;
+using SixLabors.ImageSharp.Processing.Processors.Effects;
 using SixLabors.Primitives;
 
 namespace SixLabors.ImageSharp.Processing
