+++ src/ImageSharp/Processing/Processors/Effects/PositionAwarePixelShaderProcessor{TPixel}.cs
@@ -5,36 +5,35 @@ using System;
 using System.Numerics;
 
 using SixLabors.ImageSharp.PixelFormats;
-using SixLabors.ImageSharp.Processing.Processors.PixelShading.Abstract;
 using SixLabors.Primitives;
 
-namespace SixLabors.ImageSharp.Processing.Processors.PixelShading
+namespace SixLabors.ImageSharp.Processing.Processors.Effects
 {
     /// <summary>
     /// Applies a user defined pixel shader effect through a given delegate.
     /// </summary>
     /// <typeparam name="TPixel">The pixel format.</typeparam>
-    internal class PixelShaderProcessor<TPixel> : PixelShaderProcessorBase<TPixel>
+    internal class PositionAwarePixelShaderProcessor<TPixel> : PixelShaderProcessorBase<TPixel>
         where TPixel : struct, IPixel<TPixel>
     {
         /// <summary>
         /// The user defined pixel shader.
         /// </summary>
-        private readonly PixelShader pixelShader;
+        private readonly PositionAwarePixelShader pixelShader;
 
         /// <summary>
-        /// Initializes a new instance of the <see cref="PixelShaderProcessor{TPixel}"/> class.
+        /// Initializes a new instance of the <see cref="PositionAwarePixelShaderProcessor{TPixel}"/> class.
         /// </summary>
-        /// <param name="definition">The <see cref="PixelShaderProcessor"/> defining the processor parameters.</param>
+        /// <param name="definition">The <see cref="PositionAwarePixelShaderProcessor"/> defining the processor parameters.</param>
         /// <param name="source">The source <see cref="Image{TPixel}"/> for the current processor instance.</param>
         /// <param name="sourceRectangle">The source area to process for the current processor instance.</param>
-        public PixelShaderProcessor(PixelShaderProcessor definition, Image<TPixel> source, Rectangle sourceRectangle)
+        public PositionAwarePixelShaderProcessor(PositionAwarePixelShaderProcessor definition, Image<TPixel> source, Rectangle sourceRectangle)
             : base(source, sourceRectangle)
         {
             this.pixelShader = definition.PixelShader;
         }
 
         /// <inheritdoc/>
-        protected override void ApplyPixelShader(Span<Vector4> span, int offsetY, int offsetX) => this.pixelShader(span);
+        protected override void ApplyPixelShader(Span<Vector4> span, int offsetY, int offsetX) => this.pixelShader(span, offsetY, offsetX);
     }
 }
